requirements.txt
langchain
langchain-core
langchain-google-genai
pypdf
chromadb
python-dotenv
tiktoken
sentence-transformers